+++ src/main/java/roomescape/repository/JdbcTemplateReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.persistence.jdbctemplate;
+package roomescape.repository;
 
 import java.util.List;
 import java.util.Map;
